services/dsa-service/requirements.txt
fastapi==0.111.0
uvicorn[standard]==0.30.0
motor==3.4.0
pymongo==4.6.3
python-dotenv==1.0.1
python-multipart==0.0.9
pydantic-settings==2.4.0
httpx==0.27.0
openai==1.45.0

